Place getUniquePeersbyIp() in P2P utils
