+++ elements/lisk-p2p/src/peer_selection.ts
@@ -22,28 +22,6 @@ import {
 	P2PPeerSelectionForSendInput,
 } from './p2p_types';
 
-/* tslint:disable: readonly-keyword*/
-
-export const getUniquePeersbyIp = (
-	peerList: ReadonlyArray<P2PDiscoveredPeerInfo>,
-): ReadonlyArray<P2PDiscoveredPeerInfo> => {
-	const peerMap = new Map<string, P2PDiscoveredPeerInfo>();
-
-	for (const peer of peerList) {
-		const tempPeer = peerMap.get(peer.ipAddress);
-		if (tempPeer) {
-			if (peer.height > tempPeer.height) {
-				peerMap.set(peer.ipAddress, peer);
-			}
-		} else {
-			peerMap.set(peer.ipAddress, peer);
-		}
-	}
-
-	return [...peerMap.values()];
-};
-
-/* tslint:enable: readonly-keyword */
 export const selectPeersForRequest = (
 	input: P2PPeerSelectionForRequestInput,
 ): ReadonlyArray<P2PDiscoveredPeerInfo> => {
